refactor: 未使用モデルおよび旧D&Dコンポーネントを削除
app.tsxのエントリーポイント（局員一覧・タスク一覧・シフト表リスト）から
到達できないモデルと旧D&Dコンポーネントを一掃する。

削除モデル: Assignment/AssignmentReason/ConstraintChecker, Event/SkillDefinition,
  Role, TimeSlot x2, ShiftAssignment, ShiftMatcher, ShiftEvaluation,
  MemberAssignmentEvaluation, SlotTaskEvaluation
削除コンポーネント: ShiftPlanEditor, MemberGanttEditor, MemberShiftTable,
  AssignmentEvaluation (feature/ui 各4件)
ShiftPlan.ts から deprecated な assignments 系コードを除去し、
GanttConfig/RowAvailability 型を ganttTypes.ts に集約。
