+++ shift-puzzle-bubly/shift-puzzle-libs/src/feature/index.ts
@@ -1,4 +1,3 @@
-// ---- エントリーポイントから到達可能 ----
 export * from "./MemberCollection.js";
 export * from "./MemberFilter.js";
 export * from "./MemberDetail.js";
@@ -11,9 +10,3 @@ export * from "./TaskCollection.js";
 export * from "./TaskFilter.js";
 export * from "./TaskDetail.js";
 export * from "./ShiftPlanWorldLineGraphView.js";
-
-// ---- 未使用（shift-plan-list / member-list / task-list から到達不可） ----
-export * from "./ShiftPlanEditor.js";
-export * from "./MemberShiftTable.js";
-export * from "./MemberGanttEditor.js";
-export * from "./AssignmentEvaluation.js";
