+++ shift-puzzle-bubly/shift-puzzle-libs/src/object-type-registration.ts
@@ -2,65 +2,14 @@
  * shift-puzzle 固有のオブジェクト型をレジストリに登録する
  */
 import { registerObjectType, registerObjectBubble } from "@bublys-org/bubbles-ui";
-import { getCurrentStore } from "@bublys-org/state-management";
-import {
-  ShiftPlan,
-  type ShiftPlanState,
-  type MemberState,
-  Member,
-} from "@bublys-org/shift-puzzle-model";
-import { createDefaultShifts } from "./data/sampleData.js";
 import PersonIcon from "@mui/icons-material/Person";
 import EventAvailableIcon from "@mui/icons-material/EventAvailable";
-import AssignmentIcon from "@mui/icons-material/Assignment";
 import TaskIcon from "@mui/icons-material/Task";
 import CalendarViewWeekIcon from "@mui/icons-material/CalendarViewWeek";
 import React from "react";
 
-// マスターデータ（静的）
-const shifts = createDefaultShifts();
-
-/**
- * ShiftAssignment の ID からラベルを解決する
- */
-const resolveShiftAssignmentLabel = (assignmentId: string): string | undefined => {
-  const store = getCurrentStore();
-  if (!store) return undefined;
-
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  const state = store.getState() as any;
-
-  // シフト案からドメインオブジェクトを構築して配置を検索
-  const shiftPlanStates: ShiftPlanState[] = state.shiftPlan?.shiftPlans ?? [];
-  const plans = shiftPlanStates.map(s => new ShiftPlan(s));
-
-  const assignment = plans
-    .flatMap(plan => plan.assignments)
-    .find(a => a.id === assignmentId);
-  if (!assignment) return undefined;
-
-  const shift = shifts.find(s => s.id === assignment.shiftId);
-
-  // 局員名を解決
-  const memberStateList: MemberState[] = state.shiftPuzzle?.memberList ?? [];
-  const member = memberStateList
-    .map(s => new Member(s))
-    .find(m => m.id === assignment.staffId);
-  const memberName = member?.name ?? "不明";
-
-  const shiftLabel = shift
-    ? `${shift.taskName ?? shift.taskId}(${shift.startTime}–${shift.endTime})`
-    : assignment.shiftId;
-
-  return `${shiftLabel}←${memberName}`;
-};
-
 registerObjectType('Member', React.createElement(PersonIcon, { fontSize: 'small' }));
 registerObjectType('MemberAvailability', React.createElement(EventAvailableIcon, { fontSize: 'small' }));
-registerObjectType('ShiftAssignment', {
-  icon: React.createElement(AssignmentIcon, { fontSize: 'small' }),
-  labelResolver: resolveShiftAssignmentLabel,
-});
 registerObjectType('Task', React.createElement(TaskIcon, { fontSize: 'small' }));
 registerObjectType('Shift', React.createElement(CalendarViewWeekIcon, { fontSize: 'small' }));
 
@@ -68,4 +17,3 @@ registerObjectBubble('Member',             { openingPosition: 'bubble-side' });
 registerObjectBubble('Task',               { openingPosition: 'bubble-side' });
 registerObjectBubble('MemberAvailability', { openingPosition: 'bubble-side' });
 registerObjectBubble('ShiftStatus',        { openingPosition: 'origin-side' });
-registerObjectBubble('ShiftAssignment',    { openingPosition: 'origin-side' });
