+++ shift-puzzle-bubly/shift-puzzle-libs/src/ui/index.ts
@@ -1,4 +1,3 @@
-// ---- エントリーポイントから到達可能 ----
 export * from "./MemberListView.js";
 export * from "./MemberDetailView.js";
 export * from "./MemberAvailabilityView.js";
@@ -11,9 +10,3 @@ export * from "./PrimitiveGanttView.js";
 export * from "./TaskGanttView.js";
 export * from "./TaskListView.js";
 export * from "./TaskDetailView.js";
-
-// ---- 未使用（shift-plan-list / member-list / task-list から到達不可） ----
-export * from "./MemberGanttView.js";
-export * from "./ShiftPlanTableView.js";
-export * from "./MemberShiftTableView.js";
-export * from "./AssignmentEvaluationView.js";
