--- shift-puzzle-bubly/shift-puzzle-model/src/lib/shift-plan/SlotTaskEvaluation.ts
@@ -1,4 +1,0 @@
-/**
- * @deprecated ShiftEvaluation に統合されました。このファイルは削除予定です。
- */
-export * from './ShiftEvaluation.js';
